Revert "Merge branch 'project-mvp-back' into 'dev'"

See merge request ai_track/class_07/web_project/team05/portfolio-share-service-racer!19

+++ back/src/routers/projectRouter.js
@@ -1,13 +1,13 @@
 import is from "@sindresorhus/is";
 import { Router } from "express";
 import { login_required } from "../middlewares/login_required";
-import { ProjectService } from "../services/projectService";
+import { ProjectService } from "../services/ProjectService";
 
 const projectRouter = Router();
-projectRouter.use(login_required);
 
-projectRouter.post(
-    "/project/:user_id", 
+//프로젝트 추가
+projectRouter('/projects/:user_id')
+  .post(login_required, 
     async function (req, res, next) {
     try {
     if (is.emptyObject(req.body)) {
@@ -16,12 +16,10 @@ projectRouter.post(
       );
     }
 
-    // req (request) 에서 데이터 가져오기
     const user_id = req.params.user_id;
     const title = req.body.title;
     const description = req.body.description;
 
-    // 위 데이터를 유저 db에 추가하기
     const newProject = await ProjectService.addProject({
       user_id,
       title,
@@ -37,61 +35,74 @@ projectRouter.post(
   } catch (error) {
     next(error);
   }
-});
+})
 
-
-projectRouter.get("/project/:user_id", 
-    async function(req, res, next) {
-    try {
-        const user_id = req.params.user_id
-        const project_id = req.body.project_id;
-        
-        const project = await ProjectService.getProjects({ user_id, project_id });
-
-        res.status(200).send(project);
-        } catch (error) {
-        next(error);
-        }
-    });
-
-
-
-projectRouter.put('project/:user_id', 
-    async function(req, res, next) {
+//유저의 모든 프로젝트 조회
+  .get(async function(req, res, next) {
   try {
-    const project_id = req.body.project_id;
-    const title = req.body.title ?? null;
-    const description = req.body.description ?? null;
+    const user_id = req.params.user_id;
+      
+      const project = await ProjectService.getProject({ user_id });
 
-    const toUpdate = { title, description }
+      res.status(200).send(project);
+      } catch (error) {
+      next(error);
+      }
+  })
+  
+//프로젝트 수정
+.put(login_required, 
+async function(req, res, next) {
+  try {
+      const user_id = req.params.user_id;
+      const project_id = req.body.project_id;
+      const title = req.body.title ?? null;
+      const description = req.body.description ?? null;
 
-    const updatedProject = await ProjectService.setProject({ project_id, toUpdate });
+      const toUpdate = { title, description }
 
-    if (updatedProject.errorMessage) {
-      throw new Error(updatedProject.errorMessage);
+      const updatedProject = await ProjectService.setProject({ user_id, project_id, toUpdate });
+
+      res.status(200).send(updatedProject);
+    } catch (error) {
+      next(error);
     }
-    res.status(200).send(updatedProject);
+  })
+
+.delete(login_required, 
+  async (req, res, next) => {
+  try {
+    const user_id = req.params.user_id;
+    const project_id = req.body.project_id;
+
+    const deleteResult = await ProjectService.deleteProject({ user_id, project_id });
+
+    if (!deleteResult) {
+      throw new Error("해당 프로젝트를 삭제할 수 없습니다.");
+    }
+    
+    //status 204 : 삭제요청 완료, 추가 정보없음?
+    res.status(204).send();
   } catch (error) {
     next(error);
   }
 });
 
-projectRouter.delete("/project/:user_id", 
-    async (req, res, next) => {
+
+projectRouter.get("/projects/:user_id/:project_id",
+    login_required,
+    async function (req, res, next) {
     try {
-      const project_id = req.body.project_id;
-      const deleteResult = await ProjectService.deleteProject({ project_id });
-  
-      if (!deleteResult) {
-        throw new Error("해당 프로젝트를 삭제할 수 없습니다.");
-      }
-      
-      //status 204 : 삭제요청 완료, 추가 정보없음?
-      res.status(204).send();
+      const {user_id, project_id} = req.params;
+      const projectList = await ProjectService.getProjects({ user_id });
+
+      res.status(200).send(projectList);
     } catch (error) {
       next(error);
     }
   });
-
-
+  
 export { projectRouter };
+
+  
+
